% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Crossover operator function}
\usage{
crossover(parent1, parent2, input, output, id_start)
}
\arguments{
\item{parent1}{first parent that participates in the crossover process}

\item{parent2}{second parent that participates in the crossover process}

\item{input}{number of inputs within the dataset}

\item{output}{number of outputs within the dataset}

\item{id_start}{last id non-used value}
}
\value{
Dataframe which contains two rows, one for each child
}
\description{
Crossover operator function
}
