% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Evaluation operator function}
\usage{
evaluation(individual, data, input, output, train_mode, epochs, batch_size,
  seed)
}
\arguments{
\item{individual}{individual which is evaluated}

\item{data}{data for solving both regression and classification problem}

\item{input}{number of inputs within the dataset}

\item{output}{number of outputs within the dataset}

\item{train_mode}{mode of the training. 0: multi-class, 1: regression}

\item{epochs}{number of epochs}

\item{batch_size}{size of the batch}

\item{seed}{seed}
}
\value{
Dataframe which contains the evaluated individual
}
\description{
Evaluation operator function
}
