% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialization}
\alias{initialization}
\title{Initialization operator function}
\usage{
initialization(population_size, input, output, id_start, seed)
}
\arguments{
\item{population_size}{size of the population}

\item{input}{number of inputs within the dataset}

\item{output}{number of outputs within the dataset}

\item{id_start}{last id non-used value}

\item{seed}{seed}
}
\value{
List with both individual and id.
}
\description{
Initialization operator function
}
