% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement.R
\name{replacement}
\alias{replacement}
\title{Replacement operator function}
\usage{
replacement(population, population_size)
}
\arguments{
\item{population}{population with both base population and generated childs}

\item{population_size}{size of the population}
}
\value{
Dataframe where only the best first population_size individuals are
  chosen
}
\description{
Replacement operator function
}
