% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection}
\alias{selection}
\title{Selection operator function}
\usage{
selection(population, children_number)
}
\arguments{
\item{population}{population of individuals}

\item{children_number}{number of children}
}
\value{
Dataframe with children_number rows, each one of a child
}
\description{
Selection operator function
}
